requirements.txt
attrs==21.2.0
iniconfig==1.1.1
numpy==1.21.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
toml==0.10.2
